# docs/openapi.yml
openapi: 3.1.0
info:
  title: Tracecat API
  summary: Tracecat API
  description: Tracecat is the open source Tines / Splunk SOAR alternative. You can
    operate Tracecat in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Tracecat Founders
    email: founders@tracecat.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /workflow-executions:
    get:
      tags:
      - workflow-executions
      summary: List Workflow Executions
      description: List all workflow executions.
      operationId: workflow-executions-list_workflow_executions
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: trigger
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            uniqueItems: true
            items:
              $ref: '#/components/schemas/TriggerType'
          - type: 'null'
          title: Trigger
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid4
          - $ref: '#/components/schemas/SpecialUserID'
          - type: 'null'
          title: User Id
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Limit
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: workflow_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          - type: 'null'
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowExecutionReadMinimal'
                title: Response Workflow-Executions-List Workflow Executions
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - workflow-executions
      summary: Create Workflow Execution
      description: Create and schedule a workflow execution.
      operationId: workflow-executions-create_workflow_execution
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowExecutionCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workflow-executions/{execution_id}:
    get:
      tags:
      - workflow-executions
      summary: Get Workflow Execution
      description: Get a workflow execution.
      operationId: workflow-executions-get_workflow_execution
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          title: Execution Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workflow-executions/{execution_id}/compact:
    get:
      tags:
      - workflow-executions
      summary: Get Workflow Execution Compact
      description: Get a workflow execution.
      operationId: workflow-executions-get_workflow_execution_compact
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          title: Execution Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionReadCompact'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workflow-executions/{execution_id}/cancel:
    post:
      tags:
      - workflow-executions
      summary: Cancel Workflow Execution
      description: Get a workflow execution.
      operationId: workflow-executions-cancel_workflow_execution
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          title: Execution Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workflow-executions/{execution_id}/terminate:
    post:
      tags:
      - workflow-executions
      summary: Terminate Workflow Execution
      description: Get a workflow execution.
      operationId: workflow-executions-terminate_workflow_execution
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          title: Execution Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowExecutionTerminate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EventFailure:
      properties:
        message:
          type: string
          title: Message
        cause:
          anyOf:
          - type: object
          - type: 'null'
          title: Cause
      type: object
      required:
      - message
      title: EventFailure
    EventGroup:
      properties:
        event_id:
          type: integer
          title: Event Id
        udf_namespace:
          type: string
          title: Udf Namespace
        udf_name:
          type: string
          title: Udf Name
        udf_key:
          type: string
          title: Udf Key
        action_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Id
        action_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Ref
        action_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Title
        action_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Description
        action_input:
          anyOf:
          - $ref: '#/components/schemas/RunActionInput'
          - $ref: '#/components/schemas/DSLRunArgs'
          - $ref: '#/components/schemas/GetWorkflowDefinitionActivityInputs'
          - $ref: '#/components/schemas/InteractionResult'
          - $ref: '#/components/schemas/InteractionInput'
          title: Action Input
        action_result:
          anyOf:
          - {}
          - type: 'null'
          title: Action Result
        current_attempt:
          anyOf:
          - type: integer
          - type: 'null'
          title: Current Attempt
        retry_policy:
          $ref: '#/components/schemas/ActionRetryPolicy'
        start_delay:
          type: number
          title: Start Delay
          default: 0.0
        join_strategy:
          $ref: '#/components/schemas/JoinStrategy'
          default: all
        related_wf_exec_id:
          anyOf:
          - type: string
            pattern: (?P<workflow_id>wf-[0-9a-f]{32}|wf_[0-9a-zA-Z]+)[:/](?P<execution_id>(exec_[0-9a-zA-Z]+|exec-[\w-]+|sch-[0-9a-f]{32}-.*))
          - type: 'null'
          title: Related Wf Exec Id
      type: object
      required:
      - event_id
      - udf_namespace
      - udf_name
      - udf_key
      - action_input
      title: EventGroup
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InteractionState:
      properties:
        type:
          $ref: '#/components/schemas/InteractionType'
        action_ref:
          type: string
          title: Action Ref
        status:
          $ref: '#/components/schemas/InteractionStatus'
        data:
          type: object
          title: Data
      type: object
      required:
      - type
      - action_ref
      - status
      title: InteractionState
    InteractionStatus:
      type: string
      enum:
      - idle
      - pending
      - completed
      title: InteractionStatus
    InteractionType:
      type: string
      enum:
      - approval
      - response
      title: InteractionType
    JoinStrategy:
      type: string
      enum:
      - any
      - all
      title: JoinStrategy
...
    Role:
      properties:
        type:
          type: string
          enum:
          - user
          - service
          title: Type
        workspace_id:
          anyOf:
          - type: string
            format: uuid4
          - type: 'null'
          title: Workspace Id
        user_id:
          anyOf:
          - type: string
            format: uuid4
          - type: 'null'
          title: User Id
        access_level:
          $ref: '#/components/schemas/AccessLevel'
          default: 0
        service_id:
          type: string
          enum:
          - tracecat-runner
          - tracecat-api
          - tracecat-cli
          - tracecat-schedule-runner
          - tracecat-service
          - tracecat-executor
          - tracecat-bootstrap
          title: Service Id
      type: object
      required:
      - type
      - service_id
      title: Role
      description: "The identity and authorization of a user or service.\n\nParams\n\
        ------\ntype : Literal[\"user\", \"service\"]\n    The type of role.\nuser_id\
        \ : UUID | None\n    The user's ID, or the service's user_id.\n    This can\
        \ be None for internal services, or when a user hasn't been set for the role.\n\
        service_id : str | None = None\n    The service's role name, or None if the\
        \ role is a user.\n\n\nUser roles\n----------\n- User roles are authenticated\
        \ via JWT.\n- The `user_id` is the user's JWT 'sub' claim.\n- User roles do\
        \ not have an associated `service_id`, this must be None.\n\nService roles\n\
        -------------\n- Service roles are authenticated via API key.\n- Used for\
        \ internal services to authenticate with the API.\n- A service's `user_id`\
        \ is the user it's acting on behalf of. This can be None for internal services."
...
    SpecialUserID:
      type: string
      enum:
      - current
      title: SpecialUserID
      description: A sentinel user ID that represents the current user.
...
    TriggerType:
      type: string
      enum:
      - manual
      - scheduled
      - webhook
      title: TriggerType
      description: Trigger type for a workflow execution.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WorkflowEventType:
      type: string
      enum:
      - WORKFLOW_EXECUTION_STARTED
      - WORKFLOW_EXECUTION_COMPLETED
      - WORKFLOW_EXECUTION_FAILED
      - WORKFLOW_EXECUTION_TERMINATED
      - WORKFLOW_EXECUTION_CANCELED
      - WORKFLOW_EXECUTION_CONTINUED_AS_NEW
      - WORKFLOW_EXECUTION_TIMED_OUT
      - ACTIVITY_TASK_SCHEDULED
      - ACTIVITY_TASK_STARTED
      - ACTIVITY_TASK_COMPLETED
      - ACTIVITY_TASK_FAILED
      - ACTIVITY_TASK_TIMED_OUT
      - ACTIVITY_TASK_CANCELED
      - CHILD_WORKFLOW_EXECUTION_STARTED
      - CHILD_WORKFLOW_EXECUTION_COMPLETED
      - CHILD_WORKFLOW_EXECUTION_FAILED
      - CHILD_WORKFLOW_EXECUTION_CANCELED
      - CHILD_WORKFLOW_EXECUTION_TERMINATED
      - START_CHILD_WORKFLOW_EXECUTION_INITIATED
      - CHILD_WORKFLOW_EXECUTION_TIMED_OUT
      - WORKFLOW_EXECUTION_SIGNALED
      - WORKFLOW_EXECUTION_UPDATE_ACCEPTED
      - WORKFLOW_EXECUTION_UPDATE_REJECTED
      - WORKFLOW_EXECUTION_UPDATE_COMPLETED
      title: WorkflowEventType
      description: The event types we care about.
    WorkflowExecutionCreate:
      properties:
        workflow_id:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
        inputs:
          anyOf:
          - {}
          - type: 'null'
          title: Inputs
      type: object
      required:
      - workflow_id
      title: WorkflowExecutionCreate
    WorkflowExecutionCreateResponse:
      properties:
        message:
          type: string
          title: Message
        wf_id:
          type: string
          title: Wf Id
        wf_exec_id:
          type: string
          pattern: (?P<workflow_id>wf-[0-9a-f]{32}|wf_[0-9a-zA-Z]+)[:/](?P<execution_id>(exec_[0-9a-zA-Z]+|exec-[\w-]+|sch-[0-9a-f]{32}-.*))
          title: Wf Exec Id
        payload:
          title: Payload
      type: object
      required:
      - message
      - wf_id
      - wf_exec_id
      title: WorkflowExecutionCreateResponse
    WorkflowExecutionEvent:
      properties:
        event_id:
          type: integer
          title: Event Id
        event_time:
          type: string
          format: date-time
          title: Event Time
        event_type:
          $ref: '#/components/schemas/WorkflowEventType'
        task_id:
          type: integer
          title: Task Id
        event_group:
          anyOf:
          - $ref: '#/components/schemas/EventGroup'
          - type: 'null'
          description: The action group of the event. We use this to keep track of
            what events are related to each other.
        failure:
          anyOf:
          - $ref: '#/components/schemas/EventFailure'
          - type: 'null'
        result:
          anyOf:
          - {}
          - type: 'null'
          title: Result
        role:
          anyOf:
          - $ref: '#/components/schemas/Role'
          - type: 'null'
        parent_wf_exec_id:
          anyOf:
          - type: string
            pattern: (?P<workflow_id>wf-[0-9a-f]{32}|wf_[0-9a-zA-Z]+)[:/](?P<execution_id>(exec_[0-9a-zA-Z]+|exec-[\w-]+|sch-[0-9a-f]{32}-.*))
          - type: 'null'
          title: Parent Wf Exec Id
        workflow_timeout:
          anyOf:
          - type: number
          - type: 'null'
          title: Workflow Timeout
      type: object
      required:
      - event_id
      - event_time
      - event_type
      - task_id
      title: WorkflowExecutionEvent
    WorkflowExecutionEventCompact:
      properties:
        source_event_id:
          type: integer
          title: Source Event Id
        schedule_time:
          type: string
          format: date-time
          title: Schedule Time
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
        close_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Close Time
        curr_event_type:
          $ref: '#/components/schemas/WorkflowEventType'
        status:
          $ref: '#/components/schemas/WorkflowExecutionEventStatus'
        action_name:
          type: string
          title: Action Name
        action_ref:
          type: string
          title: Action Ref
        action_input:
          anyOf:
          - {}
          - type: 'null'
          title: Action Input
        action_result:
          anyOf:
          - {}
          - type: 'null'
          title: Action Result
        action_error:
          anyOf:
          - $ref: '#/components/schemas/EventFailure'
          - type: 'null'
        child_wf_exec_id:
          anyOf:
          - type: string
            pattern: (?P<workflow_id>wf-[0-9a-f]{32}|wf_[0-9a-zA-Z]+)[:/](?P<execution_id>(exec_[0-9a-zA-Z]+|exec-[\w-]+|sch-[0-9a-f]{32}-.*))
          - type: 'null'
          title: Child Wf Exec Id
        child_wf_count:
          type: integer
          title: Child Wf Count
          default: 0
        loop_index:
          anyOf:
          - type: integer
          - type: 'null'
          title: Loop Index
      type: object
      required:
      - source_event_id
      - schedule_time
      - curr_event_type
      - status
      - action_name
      - action_ref
      title: WorkflowExecutionEventCompact
      description: A compact representation of a workflow execution event.
    WorkflowExecutionEventStatus:
      type: string
      enum:
      - SCHEDULED
      - STARTED
      - COMPLETED
      - FAILED
      - CANCELED
      - TERMINATED
      - TIMED_OUT
      - UNKNOWN
      title: WorkflowExecutionEventStatus
    WorkflowExecutionRead:
      properties:
        id:
          type: string
          title: Id
          description: The ID of the workflow execution
        run_id:
          type: string
          title: Run Id
          description: The run ID of the workflow execution
        start_time:
          type: string
          format: date-time
          title: Start Time
          description: The start time of the workflow execution
        execution_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Execution Time
          description: When this workflow run started or should start.
        close_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Close Time
          description: When the workflow was closed if closed.
        status:
          type: string
          enum:
          - RUNNING
          - COMPLETED
          - FAILED
          - CANCELED
          - TERMINATED
          - CONTINUED_AS_NEW
          - TIMED_OUT
        workflow_type:
          type: string
          title: Workflow Type
        task_queue:
          type: string
          title: Task Queue
        history_length:
          type: integer
          title: History Length
          description: Number of events in the history
        parent_wf_exec_id:
          anyOf:
          - type: string
            pattern: (?P<workflow_id>wf-[0-9a-f]{32}|wf_[0-9a-zA-Z]+)[:/](?P<execution_id>(exec_[0-9a-zA-Z]+|exec-[\w-]+|sch-[0-9a-f]{32}-.*))
          - type: 'null'
          title: Parent Wf Exec Id
        events:
          items:
            $ref: '#/components/schemas/WorkflowExecutionEvent'
          type: array
          title: Events
          description: The events in the workflow execution
        interaction_states:
          additionalProperties:
            $ref: '#/components/schemas/InteractionState'
          propertyNames:
            format: uuid
          type: object
          title: Interaction States
          description: The interactions in the workflow execution
      type: object
      required:
      - id
      - run_id
      - start_time
      - status
      - workflow_type
      - task_queue
      - history_length
      - events
      title: WorkflowExecutionRead
    WorkflowExecutionReadCompact:
      properties:
        id:
          type: string
          title: Id
          description: The ID of the workflow execution
        run_id:
          type: string
          title: Run Id
          description: The run ID of the workflow execution
        start_time:
          type: string
          format: date-time
          title: Start Time
          description: The start time of the workflow execution
        execution_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Execution Time
          description: When this workflow run started or should start.
        close_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Close Time
          description: When the workflow was closed if closed.
        status:
          type: string
          enum:
          - RUNNING
          - COMPLETED
          - FAILED
          - CANCELED
          - TERMINATED
          - CONTINUED_AS_NEW
          - TIMED_OUT
        workflow_type:
          type: string
          title: Workflow Type
        task_queue:
          type: string
          title: Task Queue
        history_length:
          type: integer
          title: History Length
          description: Number of events in the history
        parent_wf_exec_id:
          anyOf:
          - type: string
            pattern: (?P<workflow_id>wf-[0-9a-f]{32}|wf_[0-9a-zA-Z]+)[:/](?P<execution_id>(exec_[0-9a-zA-Z]+|exec-[\w-]+|sch-[0-9a-f]{32}-.*))
          - type: 'null'
          title: Parent Wf Exec Id
        events:
          items:
            $ref: '#/components/schemas/WorkflowExecutionEventCompact'
          type: array
          title: Events
          description: Compact events in the workflow execution
        interaction_states:
          additionalProperties:
            $ref: '#/components/schemas/InteractionState'
          propertyNames:
            format: uuid
          type: object
          title: Interaction States
          description: The interactions in the workflow execution
      type: object
      required:
      - id
      - run_id
      - start_time
      - status
      - workflow_type
      - task_queue
      - history_length
      - events
      title: WorkflowExecutionReadCompact
    WorkflowExecutionReadMinimal:
      properties:
        id:
          type: string
          title: Id
          description: The ID of the workflow execution
        run_id:
          type: string
          title: Run Id
          description: The run ID of the workflow execution
        start_time:
          type: string
          format: date-time
          title: Start Time
          description: The start time of the workflow execution
        execution_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Execution Time
          description: When this workflow run started or should start.
        close_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Close Time
          description: When the workflow was closed if closed.
        status:
          type: string
          enum:
          - RUNNING
          - COMPLETED
          - FAILED
          - CANCELED
          - TERMINATED
          - CONTINUED_AS_NEW
          - TIMED_OUT
        workflow_type:
          type: string
          title: Workflow Type
        task_queue:
          type: string
          title: Task Queue
        history_length:
          type: integer
          title: History Length
          description: Number of events in the history
        parent_wf_exec_id:
          anyOf:
          - type: string
            pattern: (?P<workflow_id>wf-[0-9a-f]{32}|wf_[0-9a-zA-Z]+)[:/](?P<execution_id>(exec_[0-9a-zA-Z]+|exec-[\w-]+|sch-[0-9a-f]{32}-.*))
          - type: 'null'
          title: Parent Wf Exec Id
      type: object
      required:
      - id
      - run_id
      - start_time
      - status
      - workflow_type
      - task_queue
      - history_length
      title: WorkflowExecutionReadMinimal
    WorkflowExecutionTerminate:
      properties:
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      title: WorkflowExecutionTerminate
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-tracecat-service-key
...